+++ src/main/java/is/hi/hbv501g/h6/hugboverkefni/post/PostConfig.java
@@ -26,7 +26,7 @@ public class PostConfig {
                     content,
                     user,
                     new ArrayList<Voter>(),
-                    new ArrayList<Integer>(),
+                    new ArrayList<Long>(),
                     date
             );
             userRepository.save(user);
